Reorganize app structure base on cobra docs
