+++ cmd/pull/pull.go
@@ -1,0 +1,29 @@
+package pull
+
+import (
+	"fmt"
+
+	"github.com/rocajuanma/anvil/pkg/constants"
+	"github.com/spf13/cobra"
+)
+
+var PullCmd = &cobra.Command{
+	Use:   "pull",
+	Short: "Pull assets from Github",
+	Long:  constants.PULL_COMMAND_LONG_DESCRIPTION,
+	Run: func(cmd *cobra.Command, args []string) {
+		fmt.Println("pull called")
+	},
+}
+
+func init() {
+	// Here you will define your flags and configuration settings.
+
+	// Cobra supports Persistent Flags which will work for this command
+	// and all subcommands, e.g.:
+	// setupCmd.PersistentFlags().String("foo", "", "A help for foo")
+
+	// Cobra supports local flags which will only run when this command
+	// is called directly, e.g.:
+	// setupCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+}
